Use X-FORWARDED-FOR when reverse proxied
Use the IP from the headers if the X-FORWARDED-FOR is present, which is
usually the case when geoip is being reverse proxied, i.e. when doing
TLS termination elsewhere.

Without this, geoip will try to resolve the IP address from the upstream
proxy (private in a docker network for example) instead of the HTTP
client (user's browser).